+++ CHANGELOG.rst
@@ -14,7 +14,7 @@ Version 0.4
 -----------
 | **Date:** September 10, 2020
 | **Release:** v0.4
-| **Version of CHAOS:** CHAOS-7.3 (0703)
+| **Version of CHAOS:** CHAOS-7.3 (0703), CHAOS-7.4 (0704)
 
 Features
 ^^^^^^^^